new-backend/pyproject.toml
[project]
name = "new-backend"
version = "0.1.0"
description = "FastAPI backend for SCAI workflow with LangChain + LangGraph"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.129.0",
    "langchain-core>=0.3.79",
    "langchain-community>=0.3.31",
    "langgraph>=0.4.0",
    "pandas>=2.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "snowflake-connector-python>=3.18.0",
    "snowflake-snowpark-python>=1.41.0",
    "uvicorn[standard]>=0.40.0",
    "websockets>=15.0",
    "pywinpty>=2.0.0",
]
